remove tests + cleanup
